backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.28
httpx==0.27.0
email-validator==2.1.1
